added snackbar component
