+++ app/tailwind.config.cjs
@@ -13,9 +13,10 @@ module.exports = {
         montserrat: ["var(--font-montserrat)", ...fontFamily.sans],
       },
       colors: {
-        "primary-purple": {
+        brand: {
           900: "#1C1C41",
           700: "#4E4EB2",
+          200: "#C2C2FF",
         },
         disabled: {
           400: "#C4C4CD",
